+++ R/compose.R
@@ -1,5 +1,5 @@
 #' @export
-#' @importFrom rlang eval_tidy enquo
+#' @importFrom rlang eval_tidy enquo quo_name
 #' @title Define flextable displayed values
 #' @description Modify flextable displayed values. Function is
 #' handling complex formatting as well as image insertion.
@@ -25,7 +25,7 @@
 #' @family cells formatters
 #' @section Illustrations:
 #'
-#' \if{html}{\figure{fig_compose_1.png}{options: width=80\%}}
+#' \if{html}{\figure{fig_compose_1.png}{options: width="500"}}
 compose <- function(x, i = NULL, j = NULL, value , part = "body", use_dot = FALSE){
 
   if( !inherits(x, "flextable") ) stop("compose supports only flextable objects.")
@@ -42,6 +42,11 @@ compose <- function(x, i = NULL, j = NULL, value , part = "body", use_dot = FALS
     return(x)
 
   defused_value <- enquo(value)
+  # call_label <- quo_name(defused_value)
+  # if(!grepl("as_paragraph", call_label)){
+  #   stop("argument `value` is expected to be a call to `as_paragraph()` but the value is: `", call_label, "`")
+  # }
+
   check_formula_i_and_part(i, part)
   i <- get_rows_id(x[[part]], i )
   j <- get_columns_id(x[[part]], j )