+++ R/docx_str.R
@@ -20,7 +20,7 @@ mcoalesce_options <- function(...) {
 }
 
 # docx_str -----
-docx_str <- function(x, align = "center", split = FALSE, doc = NULL, ...){
+docx_str <- function(x, align = "center", split = FALSE, keep_with_next = TRUE, doc = NULL, ...){
 
   imgs <- character(0)
   hlinks <- character(0)
@@ -62,20 +62,26 @@ docx_str <- function(x, align = "center", split = FALSE, doc = NULL, ...){
   out <- paste0(out, properties_str )
 
   if( nrow_part(x, "header") > 0 ){
-    xml_content <- format(x$header, header = TRUE, split = split, type = "wml")
+    xml_content <- format(x$header, header = TRUE,
+                          split = split, keep_with_next = keep_with_next,
+                          type = "wml")
     imgs <- append( imgs, attr(xml_content, "imgs")$image_src )
     hlinks <- append( hlinks, attr(xml_content, "htxt")$href )
     out = paste0(out, xml_content )
   }
   if( nrow_part(x, "body") > 0 ){
-    xml_content <- format(x$body, header = FALSE, split = split, type = "wml")
+    xml_content <- format(x$body, header = FALSE,
+                          split = split, keep_with_next = keep_with_next,
+                          type = "wml")
     imgs <- append( imgs, attr(xml_content, "imgs")$image_src )
     hlinks <- append( hlinks, attr(xml_content, "htxt")$href )
 
     out = paste0(out, xml_content )
   }
   if( nrow_part(x, "footer") > 0 ){
-    xml_content <- format(x$footer, header = FALSE, split = split, type = "wml")
+    xml_content <- format(x$footer, header = FALSE,
+                          split = split, keep_with_next = keep_with_next,
+                          type = "wml")
     imgs <- append( imgs, attr(xml_content, "imgs")$image_src )
     hlinks <- append( hlinks, attr(xml_content, "htxt")$href )
     out = paste0(out, xml_content )
@@ -114,7 +120,6 @@ caption_docx_bookdown <- function(x){
   tab_props$id <- mcoalesce_options(x$caption$autonum$bookmark, tab_props$id, opts_current$get('label'))
   tab_props$cap <- mcoalesce_options(x$caption$value, tab_props$cap)
   tab_props$cap.style <- mcoalesce_options(x$caption$style, tab_props$cap.style)
-  tab_props$cap.fp_text <- mcoalesce_options(x$caption$fp_text, tab_props$cap.fp_text)
 
   has_caption_label <- !is.null(tab_props$cap)
   has_caption_style <- !is.null(tab_props$cap.style)
@@ -131,29 +136,10 @@ caption_docx_bookdown <- function(x){
   caption <- tab_props$cap
 
   zz <- if(!is.null(tab_props$id)){
-    run_bookmark(tab_props$id, ftext("TABCAPTION"))
+    paste0("(\\#", tab_props$tab.lp, tab_props$id,")")
   } else {
-    ftext("TABCAPTION")
+    ""
   }
-  zz <- paste("`", to_wml(zz), "`{=openxml}", sep = "")
-
-
-  # break the xml so that pandoc manage captions
-
-  replacement <- paste0(
-    "`{=openxml}",
-    paste0(
-      if(isTRUE(tab_props$cap.fp_text$bold)) "**",
-      if(isTRUE(tab_props$cap.fp_text$italic)) "*",
-      "(\\\\#",
-      tab_props$tab.lp, tab_props$id,
-      ")",
-      if(isTRUE(tab_props$cap.fp_text$italic)) "*",
-      if(isTRUE(tab_props$cap.fp_text$bold)) "**"
-    ),
-    "`")
-
-  zz <- gsub("<w:r>(.*)</w:r>", replacement, zz)
 
   caption <- paste(
     style_start,
@@ -169,7 +155,7 @@ caption_docx_standard <- function(x){
 
   caption_label <- mcoalesce_options(x$caption$value, tab_props$cap)
   caption_style <- mcoalesce_options(x$caption$style, tab_props$cap.style)
-  caption_id <- mcoalesce_options(x$caption$autonum$bookmark, tab_props$id)
+  caption_id <- mcoalesce_options(x$caption$autonum$bookmark, tab_props$id, opts_current$get('label'))
   caption_lp <- mcoalesce_options(if( !is.null(tab_props$tab.lp) )
                                     gsub(":$", "", tab_props$tab.lp)
                                   else NULL,